Retire 3.0 image
We never really used MongoDB 3.0 (no customer MongoDB instances are
running on it), and it doesn't properly handle a SIGTERM (it never
exits), so we'd rather avoid letting customers launch new instances
using 3.0.

The underlying image will continue to exist in Quay (for now), but
there's no reason to continue making new ones.