Add --initialize-backup command
This command will be triggered by Sweetness in order to purge the local
replica set configuration when initializing from a backup.

This ensures the new database comes online in a standalone replica set,
whereas it would otherwise attempt to join the original database's
replica set (and transition to `REMOVED` in the process, since it's not
a member of that replica set).